requirements.txt
cloudformation-cli-python-lib>=2.1.3
kafka-python==2.0.2
confluent-kafka==1.6.1
cfn_resource_provider>=1.0
importlib-metadata>=3.10.0
typing-extensions>=3.7.4.3
